double 에서 long으로 상장가격 데이터 수정

+++ src/main/java/com/spike/controller/SecuritiesAccountController.java
@@ -41,7 +41,7 @@ public class SecuritiesAccountController {
 
 	@PostMapping("/open")
 	public String createAccount(
-			@RequestParam("initialDeposit") Double initialDeposit,
+			@RequestParam("initialDeposit") long initialDeposit,
 			@RequestParam("currency") String currency,
 			@RequestParam("accountPassword") String accountPassword,
 			Principal principal,
